server/requirements.txt
asgiref==3.3.1
dj-database-url==0.5.0
Django==3.1.6
django-cors-headers==3.7.0
django-environ==0.4.5
django-filter==2.4.0
djangorestframework==3.12.2
djangorestframework-simplejwt==4.6.0
gunicorn==20.0.4
Markdown==3.3.3
psycopg2-binary==2.8.6
PyJWT==2.0.1
pytz==2021.1
sqlparse==0.4.1
whitenoise==5.2.0
